Hibernate db fix
Hibernate db fix + tests up

Issue-ID: PORTAL-710

+++ portal-BE/src/main/java/org/onap/portal/service/fn/FnLuTimezoneService.java
@@ -57,7 +57,7 @@ public class FnLuTimezoneService {
               this.fnLuTimezoneDao = fnLuTimezoneDao;
        }
 
-       public Optional<FnLuTimezone> getById(Integer id){
+       public Optional<FnLuTimezone> getById(Long id){
               return fnLuTimezoneDao.findById(id);
        }
 }
